--- a/Changes
+++ b/Changes
@@ -255,6 +255,10 @@ _______________
   (Sébastien Hinderer, review by Gabriel Scherer, Antonin Décimo,
   Miod Vallat and Samuel Hym)
 
+* #13201: Fix and speedup builds with TSan.
+  (Sébastien Hinderer, review by Miod Vallat, Gabriel Scherer and
+  Olivier Nicole)
+
 ### Bug fixes:
 
 - #12854: Add a test in the regression suite that flags the bug #12825.