--- a/Makefile.build_config.in
+++ b/Makefile.build_config.in
@@ -70,20 +70,6 @@ OC_INCLUDES = $(addprefix -I $(ROOTDIR)/, \
   runtime @flexdll_source_dir@ @winpthreads_source_include_dir@)
 OC_CPPFLAGS = $(OC_INCLUDES) @oc_cppflags@
 
-# The following variable defines flags to be passed to the C preprocessor
-# when compiling C files to be linked with native code. This includes
-# the native runtime itself and can also include the stub code around
-# C libraries when it needs to be different from the one used to
-# link with bytecode.
-
-# These flags should be passed *in addition* to those in OC_CPPFLAGS, they
-# should not replace them.
-
-OC_NATIVE_CPPFLAGS=-DNATIVE_CODE @native_cppflags@
-
-# Same as above, for CFLAGS
-OC_NATIVE_CFLAGS=@native_cflags@
-
 # Additional link-time options
 # To support dynamic loading of shared libraries (they need to look at
 # our own symbols):