--- a/configure
+++ b/configure
@@ -3270,8 +3270,8 @@ oc_exe_ldflags=""
 
 tsan=false
 # Passed to the linker by ocamlopt when tsan is enabled
-oc_tsan_cflags="-fsanitize=thread"
-oc_tsan_cppflags="-DWITH_THREAD_SANITIZER"
+tsan_cflags="-fsanitize=thread"
+tsan_cppflags="-DWITH_THREAD_SANITIZER"
 tsan_distinguish_volatile_cflags=""
 
 # The C# compiler and its flags
@@ -17123,7 +17123,7 @@ rm -f core conftest.err conftest.$ac_objext conftest.beam \
   gcc-[0123456789]-*|gcc-10-*|clang-*) :
      ;; #(
   *) :
-    oc_tsan_cflags="$oc_tsan_cflags -Wno-tsan" ;;
+    tsan_cflags="$tsan_cflags -Wno-tsan" ;;
 esac
   case $ocaml_cc_vendor in #(
   gcc*) :
@@ -17173,10 +17173,10 @@ else $as_nop
   as_fn_error $? "The C compiler does not support the \`$tsan_distinguish_volatile_cflags' flag. Try upgrading to GCC >= 11, or to Clang >= 11." "$LINENO" 5
 fi
 
-  oc_tsan_cflags="$oc_tsan_cflags $tsan_distinguish_volatile_cflags"
-  common_cppflags="$common_cppflags $oc_tsan_cppflags"
-  native_cflags="$native_cflags $oc_tsan_cflags"
-  ocamlc_cflags="$ocamlc_cflags $oc_tsan_cflags"
+  tsan_cflags="$tsan_cflags $tsan_distinguish_volatile_cflags"
+  common_cppflags="$common_cppflags $tsan_cppflags"
+  native_cflags="$native_cflags $tsan_cflags"
+  ocamlc_cflags="$ocamlc_cflags $tsan_cflags"
   tsan_native_runtime_c_sources="tsan"
 else $as_nop
   { printf "%s\n" "$as_me:${as_lineno-$LINENO}: not using thread sanitizer" >&5