--- a/utils/config.common.ml.in
+++ b/utils/config.common.ml.in
@@ -77,10 +77,10 @@ let configuration_variables () =
   p "standard_library" standard_library;
   p "ccomp_type" ccomp_type;
   p "c_compiler" c_compiler;
-  p "ocamlc_cflags" ocamlc_cflags;
-  p "ocamlc_cppflags" ocamlc_cppflags;
-  p "ocamlopt_cflags" ocamlopt_cflags;
-  p "ocamlopt_cppflags" ocamlopt_cppflags;
+  p "bytecode_cflags" bytecode_cflags;
+  p "bytecode_cppflags" bytecode_cppflags;
+  p "native_cflags" native_cflags;
+  p "native_cppflags" native_cppflags;
   p "bytecomp_c_compiler" bytecomp_c_compiler;
   p "native_c_compiler" native_c_compiler;
   p "bytecomp_c_libraries" bytecomp_c_libraries;