--- a/utils/config.generated.ml.in
+++ b/utils/config.generated.ml.in
@@ -27,23 +27,24 @@ let c_compiler = {@QS@|@CC@|@QS@}
 let c_output_obj = {@QS@|@outputobj@|@QS@}
 let c_has_debug_prefix_map = @cc_has_debug_prefix_map@
 let as_has_debug_prefix_map = @as_has_debug_prefix_map@
-let ocamlc_cflags = {@QS@|@ocamlc_cflags@|@QS@}
-let ocamlc_cppflags = {@QS@|@ocamlc_cppflags@|@QS@}
-(* #7678: ocamlopt uses these only to compile .c files, and the behaviour for
-          the two drivers should be identical. *)
-let ocamlopt_cflags = {@QS@|@ocamlc_cflags@|@QS@}
-let ocamlopt_cppflags = {@QS@|@ocamlc_cppflags@|@QS@}
+let bytecode_cflags = {@QS@|@bytecode_cflags@|@QS@}
+let bytecode_cppflags = {@QS@|@bytecode_cppflags@|@QS@}
+let native_cflags =
+  {@QS@|@bytecode_cflags@|@QS@} ^ " " ^ {@QS@|@native_cflags@|@QS@}
+let native_cppflags =
+  {@QS@|@bytecode_cppflags@|@QS@} ^ " " ^ {@QS@|@native_cppflags@|@QS@}
+
 let bytecomp_c_libraries = {@QS@|@zstd_libs@ @cclibs@|@QS@}
 (* bytecomp_c_compiler and native_c_compiler have been supported for a
    long time and are retained for backwards compatibility.
    For programs that don't need compatibility with older OCaml releases
    the recommended approach is to use the constituent variables
-   c_compiler, ocamlc_cflags, ocamlc_cppflags etc., directly.
+   c_compiler, {bytecode,native}_c[pp]flags etc. directly.
 *)
 let bytecomp_c_compiler =
-  c_compiler ^ " " ^ ocamlc_cflags ^ " " ^ ocamlc_cppflags
+  c_compiler ^ " " ^ bytecode_cflags ^ " " ^ bytecode_cppflags
 let native_c_compiler =
-  c_compiler ^ " " ^ ocamlopt_cflags ^ " " ^ ocamlopt_cppflags
+  c_compiler ^ " " ^ native_cflags ^ " " ^ native_cppflags
 let native_c_libraries = {@QS@|@cclibs@|@QS@}
 let native_ldflags = {@QS@|@native_ldflags@|@QS@}
 let native_pack_linker = {@QS@|@PACKLD@|@QS@}