This PR fixes issue #13042 and also addresses the problem PR #12781 attempted to resolve.

Both of these are instances of artefacts for bytecode programs or libraries being instrumented whereas they should not be.

The present PR makes sure only the artefacts related to native programs and libraries get instrumented. As a side effect, the build time for the compiler when configured with `--endable-tsan` is considerably reduced, as can be seen from these results obtained with `make -j`:

Trunk:

```
real    6m27,551s
user    41m52,165s
sys     7m14,887s
```

PR branch:

```
real    3m18,914s
user    18m52,346s
sys     1m57,559s
```

This PR consists of the four following commits:

 1. Variable renamings local to the `configure` script.
 2. Make sure `runtime/tsan.c` is compiled and linked only for native programs and libraries.
 3. Restoring the possibility to use different C and C preprocessor flags when compiling C programs in bytecode and native contexts.
 4. Taking advantage of the previous commit to make sure only the bytecode-related artefacts get instrumented.

Of these four commits, the third one can certainly be seen as the core of this PR. It basically reverts 2b4fe09dc7196a4814710495e29a183dbeabad06, which had been submitted as part of PR #8631. However, the present PR proposes new names for the different types of flags, which are based on the backend for which we compile rather than on the name of the associated compiler, since the backend names seem more stable than the compiler names. So we use `bytecode_cflags` rather than `ocamlc_cflags`, etc., for the day when `ocamlc` and `ocamlopt` will have a unified compiler driver. ;)

This work may be of interest to @fabbing and @OlivierNicole, among others.