@shindere 2024-05-30
I will rebase yes, and intend to go through the PR again because I think
there is still room for improvemnet. In particular, I think it's
important to pass the TSan flags at link time because, as I understood
it, some platforms do require that.