@shindere 2024-05-30
I just pushed an updated version of the PR that passes CI.

(I know it does because it did and I just squashed a tentative fix in the
commit it was actually fixing.)

Basically, the problem was that there were some flags missing in those
ocamlopt must pass to the C files it compiles to be linked with native code.

This is, in turn, the consequence of two things. On the one hand, it is
currently true that the C flags used for native code are a superset of those
used for bytecode. On the other hand, this assumption, which is true
currently but may not always be true, is expressed at several places in the
build system, namely everytime we say that the native flags are the bytecode
flags plus something else.

I think this should be fixed and the build system shoudl not compute the
native code flags in terms of the bytecode ones. It should rather use the
native code flags, whatever they are, and those should be computed
once and for all during the configure stage, where we can compute
them in connection with the bytecode flags as appropriate.
But that's quite a bit of work and quite orthogonal to the current
PR so I would like to leave it for a dedicated PR which I intend to
submit soon.