@shindere 2024-06-03
Miod Vallat (2024/06/03 02:52 -0700):
> @dustanddreams commented on this pull request.
>
>
>
> > @@ -2339,6 +2340,7 @@ AS_CASE([$host],
>      [PTHREAD_LIBS=''],
>    [AX_PTHREAD(
>      [common_cflags="$common_cflags $PTHREAD_CFLAGS"
> +    AC_MSG_NOTICE([SEB: common_cflags="$common_cflags"])
>
> You forgot that leftover debug message.
> ```suggestion
> ```

Oops! Thanks! Fixed!