@shindere 2024-06-04
Thanks a lot @oliviernicole for your very thorough review.

Olivier has asked very good quesitons and I was happy to have the
possibility to explain in a ocnversation rather than in a written form,
until we hae a good documentaiton for all the build system.