@anmonteiro 2024-06-15
> So we use bytecode_cflags rather than ocamlc_cflags, etc., for the day when ocamlc and ocamlopt will have a unified compiler driver. ;)

Could someone provide a stronger argument for breaking every single user of `ocamlc -config` starting in 5.3 other than the hand-wavy "look, guys, it'll be better when the compiler driver is unified"?