@anmonteiro 2024-06-15
No sarcasm intended, this is a pretty serious regression. On OCaml 5.2:

```
$ ocamlc.opt -config
...
ocamlc_cflags:  -O2 -fno-strict-aliasing -fwrapv -pthread   -pthread
ocamlc_cppflags:  -D_FILE_OFFSET_BITS=64
ocamlopt_cflags:  -O2 -fno-strict-aliasing -fwrapv -pthread   -pthread
ocamlopt_cppflags:  -D_FILE_OFFSET_BITS=64
...
```

On OCaml trunk (what will be 5.3):

```
$ ocamlc.opt -config
...
bytecode_cflags:  -O2 -fno-strict-aliasing -fwrapv -pthread   -pthread
bytecode_cppflags:  -D_FILE_OFFSET_BITS=64
native_cflags:  -O2 -fno-strict-aliasing -fwrapv -pthread   -pthread
native_cppflags:  -D_FILE_OFFSET_BITS=64  -DNATIVE_CODE -DTARGET_arm64 -DMODEL_default -DSYS_macosx
...
```

Tooling such as dune [relies](https://github.com/ocaml/dune/blob/987e1f16b3e6a9d43ec7e3cb130fe52ec86fceca/src/ocaml-config/ocaml_config.ml#L558-L561) on those variables being present.