@anmonteiro 2024-06-15
If there's a strong desire to rename theses variables, you can still introduce the new variables alongside the current ones and establish some kind of deprecation strategy. 

That's what puzzled me so much, given the OCaml compiler's track record of not introducing breaking changes like these.