@shindere 2024-06-17
Antonio Nuno Monteiro (2024/06/14 21:43 -0700):
> If there's a strong desire to rename theses variables, you can still
> introduce the new variables alongside the current ones and establish
> some kind of deprecation strategy.

Taht was actually my initial intention, and then I forgot about it, I
apologize for that and the trouble it created.

> That's what puzzled me so much, given the OCaml compiler's track
> record of not introducing breaking changes like these.

So is it okay if I (1) submit a PR that re-introduces the older names
but with deprecation marks and (2) make sure dune is transitionning to
the new names?

I intend to submit this as part of a revised version of #12586 I am
currently working on, unless people feel there is a need for a dedicated
PR but I'd definitely prefer to have this in #12586 which basically
deals with C and CPP flags.