@shindere 2024-06-17
I'd like to insist on the renaming because it's more future proof.

As I think I already wrote, I am hoping that we will once have a unified
compiler that will have several backends, including bytecode and native.
To me the proposed names work better in such a context and are easier to
extend to potential other backends.