@shindere 2024-06-17
Am I correct that re-introducing the old names with deprecation would fix
trunk? If so I can prepare a PR with just that promptly.