@shindere 2024-06-18
Just opened #13244 to fix `ocamlc -config`